less debug spam

+++ nd4j-parameter-server-parent/nd4j-parameter-server-node/src/main/java/org/nd4j/parameterserver/distributed/messages/aggregations/BaseAggregation.java
@@ -79,7 +79,7 @@ public abstract class BaseAggregation extends BaseVoidMessage implements VoidAgg
     public INDArray getAccumulatedResult() {
 
         if (aggregationWidth == 1) {
-            return chunks.get(0);
+            return chunks.get((short) 0);
         } else return Nd4j.hstack(chunks.values());
     }
